test(api): split pure extract_tool_sources unit tests out of the integration module (PR6c)
The three pure unit tests for extract_tool_sources (no DB required) were inheriting
the module-level pytestmark = pytest.mark.integration, causing pytest -m "not integration"
to skip them incorrectly. Move them to test_extract_tool_sources.py with no integration mark.
